% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtrack-package.R
\docType{package}
\name{Rtrack-package}
\alias{Rtrack}
\alias{Rtrack-package}
\title{Rtrack: Spatial Navigation Strategy Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A toolkit for the analysis of paths from spatial tracking experiments and calculation of goal-finding strategies. This package is centered on an approach using machine learning for path classification.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rupertoverall.net/Rtrack/}
  \item \url{https://github.com/rupertoverall/Rtrack}
  \item Report bugs at \url{https://github.com/rupertoverall/Rtrack/issues}
}

}
\author{
\strong{Maintainer}: Rupert Overall \email{rtrack@rupertoverall.net} (\href{https://orcid.org/0000-0002-3882-6073}{ORCID})

}
\keyword{internal}
