% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtrack.R
\docType{package}
\name{Rtrack}
\alias{Rtrack}
\title{Analysis of spatial tracking data and strategy classification.}
\description{
Analysis of spatial tracking data and strategy classification.
}
\section{Functions provided}{

\code{\link{calculate_metrics}}

\code{\link{call_mwm_strategy_garthe}}

\code{\link{call_strategy}}

\code{\link{check_experiment}}

\code{\link{export_json}}

\code{\link{export_results}}

\code{\link{identify_track_format}}

\code{\link{plot_density}}

\code{\link{plot_path}}

\code{\link{plot_strategies}}

\code{\link{plot_variable}}

\code{\link{read_arena}}

\code{\link{read_experiment}}

\code{\link{read_path}}

\code{\link{threshold_strategies}}
}

